requirements.txt
# Core dependencies
streamlit==1.28.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0

# Data processing
pandas==2.1.3
numpy==1.24.3

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2

# Time series forecasting
statsmodels==0.14.0
tensorflow==2.15.0
keras==2.15.0

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# API requests
requests==2.31.0
aiohttp==3.9.1

# Utilities
joblib==1.3.2